// source/auth/mock-auth-types.ts
/**
 * @license Use of this source code is governed by a GPL-3.0 license that
 * can be found in the LICENSE file at https://github.com/cartant/firebase-nightlight
 */

export interface MockIdentity {
    credential?: any;
    email: string;
    password?: string;
    token?: string;
    uid?: string;
}